+++ Backend-Ecommerce/src/services/schemas/schemaSuppliers.ts
@@ -10,7 +10,7 @@ const schemaSuppliers = new Schema({
             location: {type: String, require: true}
         }
     ],
-    products: [{type: Schema.Types.ObjectId, ref: 'Product'}]
+    products: [{type: Schema.Types.ObjectId, ref: 'Product', required: true}]
 })
 
 export default schemaSuppliers